refactor: 간단한 useContext 는 프로바이더와 함께 위치
