--- src/hooks/useDictionary.tsx
@@ -1,7 +1,0 @@
-import { useContext } from 'react';
-import { DictionaryContext } from '@/context/DictionaryProvider';
-
-export const useDictionary = () => {
-  const context = useContext(DictionaryContext);
-  return context;
-};
